+++ jdempotent/src/test/java/br/com/sawcunhaos/foundation/jdempotent/core/utils/TestIdempotentResource.java
@@ -14,12 +14,18 @@
 package br.com.sawcunhaos.foundation.jdempotent.core.utils;
 
 
+import br.com.sawcunhaos.foundation.jdempotent.api.IdempotentFailurePolicy;
 import br.com.sawcunhaos.foundation.jdempotent.api.JdempotentRequestPayload;
 import br.com.sawcunhaos.foundation.jdempotent.api.JdempotentResource;
 import org.springframework.stereotype.Component;
 
+import java.util.concurrent.atomic.AtomicInteger;
+
 @Component
 public class TestIdempotentResource {
+
+    private final AtomicInteger keepFailedInvocationCount = new AtomicInteger();
+
     @JdempotentResource
     public void idempotentMethod(IdempotentTestPayload testObject) {
     }
@@ -27,6 +33,16 @@ public class TestIdempotentResource {
     @JdempotentResource(cachePrefix = "TestIdempotentResource")
     public void idempotentMethodThrowingARuntimeException(IdempotentTestPayload testObject) {
         throw new TestException();
+    }
+
+    @JdempotentResource(cachePrefix = "TestIdempotentResource", onBusinessException = IdempotentFailurePolicy.KEEP_FAILED)
+    public void idempotentMethodThrowingARuntimeExceptionKeepFailed(IdempotentTestPayload testObject) {
+        keepFailedInvocationCount.incrementAndGet();
+        throw new TestException();
+    }
+
+    public int getKeepFailedInvocationCount() {
+        return keepFailedInvocationCount.get();
     }
 
     @JdempotentResource(cachePrefix = "TestIdempotentResource")